% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runway_to_sf.R
\name{build_runway_geometry}
\alias{build_runway_geometry}
\title{Constrói geometria 3D da pista}
\usage{
build_runway_geometry(utm_data, largura_m, elev_cab1, elev_cab2)
}
\description{
Constrói geometria 3D da pista
}
\keyword{internal}
