% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runway_to_sf.R
\name{calculate_runway_distances}
\alias{calculate_runway_distances}
\title{Calcula distâncias e valida comprimentos}
\usage{
calculate_runway_distances(
  utm_data,
  comprimento_declarado,
  pista_id,
  debug,
  distance,
  cab1,
  cab2
)
}
\description{
Calcula distâncias e valida comprimentos
}
\keyword{internal}
