% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runway_strip_to_sf.R
\name{calculate_strip_geometry}
\alias{calculate_strip_geometry}
\title{Calcula geometria da faixa em coordenadas UTM}
\usage{
calculate_strip_geometry(thr1_utm, thr2_utm, ext_long, largura_total)
}
\description{
Calcula geometria da faixa em coordenadas UTM
}
\keyword{internal}
