% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runway_strip_to_sf.R
\name{convert_to_utm}
\alias{convert_to_utm}
\title{Converte cabeceiras para coordenadas UTM}
\usage{
convert_to_utm(cab1, cab2)
}
\description{
Converte cabeceiras para coordenadas UTM
}
\keyword{internal}
