% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runway_to_sf.R
\name{create_runway_metadata}
\alias{create_runway_metadata}
\title{Cria metadados da pista}
\usage{
create_runway_metadata(pista_id, pista_data, cab1, cab2, distances)
}
\description{
Cria metadados da pista
}
\keyword{internal}
