% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_runways_leaflet.R
\name{create_runway_popup_simple}
\alias{create_runway_popup_simple}
\title{Cria popup HTML simples para pista}
\usage{
create_runway_popup_simple(pista, cor)
}
\description{
Cria popup HTML simples para pista
}
\keyword{internal}
