% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_runways.R
\name{create_runway_structure}
\alias{create_runway_structure}
\title{Cria estrutura padronizada para uma pista}
\usage{
create_runway_structure(pista_data)
}
\description{
Cria estrutura padronizada para uma pista
}
\keyword{internal}
