% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runway_strip_to_sf.R
\name{create_strip_geometry_3d}
\alias{create_strip_geometry_3d}
\title{Cria geometria 3D da faixa}
\usage{
create_strip_geometry_3d(coords_utm_2d, epsg_utm, z_max, crs_final)
}
\description{
Cria geometria 3D da faixa
}
\keyword{internal}
