% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runway_strip_to_sf.R
\name{create_strip_metadata}
\alias{create_strip_metadata}
\title{Cria metadados da faixa}
\usage{
create_strip_metadata(airport, pista_id, pista_data, faixa_params, z_max)
}
\description{
Cria metadados da faixa
}
\keyword{internal}
