% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_airport_layers.R
\name{create_strip_popup}
\alias{create_strip_popup}
\title{Cria popup para faixa de pista}
\usage{
create_strip_popup(faixa, cor, pista_id)
}
\description{
Cria popup para faixa de pista
}
\keyword{internal}
