% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strip_params.R
\name{create_strip_result}
\alias{create_strip_result}
\title{Cria resultado estruturado para uma pista}
\usage{
create_strip_result(pista_id, pista_data, params)
}
\description{
Cria resultado estruturado para uma pista
}
\keyword{internal}
