% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Airport.R
\name{.calc_runway_lengths}
\alias{.calc_runway_lengths}
\title{Distâncias calculadas (UTM e haversine) com diferença vs declarado}
\usage{
.calc_runway_lengths(cab1, cab2, declarado)
}
\description{
Distâncias calculadas (UTM e haversine) com diferença vs declarado
}
\keyword{internal}
