% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strip_params.R
\name{.get_strip_params}
\alias{.get_strip_params}
\title{Calcula faixas de pista com metadados da consulta RBAC-154}
\usage{
.get_strip_params(airport)
}
\arguments{
\item{airport}{Objeto Airport}
}
\value{
Lista com parâmetros + metadados para validação
}
\description{
Calcula faixas de pista com metadados da consulta RBAC-154
}
\keyword{internal}
