% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_runways.R
\name{.process_runways}
\alias{.process_runways}
\title{Processa dados JSON de pistas para estrutura interna}
\usage{
.process_runways(pistas_json)
}
\arguments{
\item{pistas_json}{Lista de pistas do JSON do aeroporto}
}
\value{
Lista estruturada com dados das pistas
}
\description{
Processa dados JSON de pistas para estrutura interna
}
\keyword{internal}
