% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_airport_json.R
\name{.validate_airport_json}
\alias{.validate_airport_json}
\title{Valida estrutura básica de JSON de aeroporto}
\usage{
.validate_airport_json(json)
}
\arguments{
\item{json}{Lista JSON carregada}
}
\value{
Logical indicando se estrutura é válida
}
\description{
Valida estrutura básica de JSON de aeroporto
}
\keyword{internal}
