% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strip_params.R
\name{find_rbac_params}
\alias{find_rbac_params}
\title{Busca parâmetros RBAC-154 por código e tipo de aproximação}
\usage{
find_rbac_params(parametros, code_number, approach_type)
}
\description{
Busca parâmetros RBAC-154 por código e tipo de aproximação
}
\keyword{internal}
