% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_superficie_params.R
\name{list_superficies}
\alias{list_superficies}
\title{Lista todas as cabeceiras com suas superfícies ICA 11-408 – FORMATO RELATÓRIO OFICIAL}
\usage{
list_superficies(apt)
}
\arguments{
\item{apt}{Objeto Airport}
}
\description{
Lista todas as cabeceiras com suas superfícies ICA 11-408 – FORMATO RELATÓRIO OFICIAL
}
