% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_airport_layers.R
\name{plot_airport_layers}
\alias{plot_airport_layers}
\title{Plota todas as camadas de um aeroporto de forma otimizada}
\usage{
plot_airport_layers(
  airport,
  incluir_faixas = TRUE,
  cores_pistas = NULL,
  cores_faixas = NULL,
  opacidade_pistas = 0.9,
  opacidade_faixas = 0.4
)
}
\arguments{
\item{airport}{Objeto Airport}

\item{incluir_faixas}{Lógico. Incluir faixas? (padrão: TRUE)}

\item{cores_pistas}{Vetor de cores para pistas ou NULL para automático}

\item{cores_faixas}{Vetor de cores para faixas ou NULL para automático}

\item{opacidade_pistas}{Opacidade das pistas (0-1)}

\item{opacidade_faixas}{Opacidade das faixas (0-1)}
}
\value{
Objeto leaflet
}
\description{
Plota todas as camadas de um aeroporto de forma otimizada
}
