% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Airport.R
\name{print.Airport}
\alias{print.Airport}
\title{Método print personalizado para objetos Airport – FORMATO RELATÓRIO OFICIAL}
\usage{
\method{print}{Airport}(x, ...)
}
\arguments{
\item{x}{Objeto Airport}

\item{...}{Argumentos adicionais}
}
\value{
Invisibly retorna o objeto x
}
\description{
Método print personalizado para objetos Airport – FORMATO RELATÓRIO OFICIAL
}
