% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_superficies.R
\name{print.superficies}
\alias{print.superficies}
\title{Relatório completo ICA 11-408 – TODOS OS TEXTOS EM PRETO}
\usage{
\method{print}{superficies}(apt)
}
\description{
Relatório completo ICA 11-408 – TODOS OS TEXTOS EM PRETO
}
