% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runway_to_sf.R
\name{print_runway_debug}
\alias{print_runway_debug}
\title{Imprime diagnóstico da pista}
\usage{
print_runway_debug(
  pista_id,
  declarado,
  distancia_utm,
  diferenca_utm,
  distancia_haversine = NA_real_,
  diferenca_haversine = NA_real_
)
}
\description{
Imprime diagnóstico da pista
}
\keyword{internal}
