% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_runways.R
\name{process_runway_thresholds}
\alias{process_runway_thresholds}
\title{Processa dados das cabeceiras}
\usage{
process_runway_thresholds(cabeceiras_json)
}
\description{
Processa dados das cabeceiras
}
\keyword{internal}
