% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runway_to_sf.R
\name{transform_runway_geometry}
\alias{transform_runway_geometry}
\title{Converte geometria para sistema de coordenadas final}
\usage{
transform_runway_geometry(runway_geom, epsg_utm, proj)
}
\description{
Converte geometria para sistema de coordenadas final
}
\keyword{internal}
